move wfs and wms types & interfaces to separate files
