+++ src/index.ts
@@ -1,5 +1,5 @@
+export { default as WfsEndpoint } from './wfs/endpoint';
 export {
-  default as WfsEndpoint,
   WfsVersion,
   WfsFeatureWithProps,
   WfsFeatureTypeSummary,
@@ -10,15 +10,15 @@ export {
   WfsFeatureTypePropDetails,
   WfsFeatureTypePropsDetails,
   WfsFeatureTypeUniqueValue,
-} from './wfs/endpoint';
+} from './wfs/model';
+export { default as WmsEndpoint } from './wms/endpoint';
 export {
-  default as WmsEndpoint,
   WmsLayerFull,
   WmsVersion,
   WmsLayerSummary,
   LayerStyle,
   LayerAttribution,
-} from './wms/endpoint';
+} from './wms/model';
 export { default as WmtsEndpoint } from './wmts/endpoint';
 export {
   LayerDimensionValue,
